+++ src/main/groovy/nebula/plugin/dependencylock/DependencyLockPlugin.groovy
@@ -35,69 +35,98 @@ class DependencyLockPlugin implements Plugin<Project> {
         String clLockFileName = project.hasProperty('dependencyLock.lockFile') ? project['dependencyLock.lockFile'] : null
         DependencyLockExtension extension = project.extensions.create('dependencyLock', DependencyLockExtension)
 
-        GenerateLockTask lockTask = project.tasks.create('generateLock', GenerateLockTask)
-        lockTask.conventionMapping.with {
-            dependenciesLock = {
-                new File(project.buildDir, clLockFileName ?: extension.lockFile)
-            }
-            configurationNames = { extension.configurationNames }
-        }
-
-        SaveLockTask saveTask = project.tasks.create('saveLock', SaveLockTask)
-        saveTask.conventionMapping.with {
-            generatedLock = { lockTask.dependenciesLock }
-            outputLock = { new File(project.projectDir, extension.lockFile) }
-        }
-        saveTask.dependsOn lockTask
-
         Map overrides = loadOverrides()
+        GenerateLockTask lockTask = configureLockTask(clLockFileName, extension, overrides)
+        configureSaveTask(lockTask, extension)
 
         project.gradle.taskGraph.whenReady { taskGraph ->
             File dependenciesLock = new File(project.projectDir, clLockFileName ?: extension.lockFile)
 
             if (!taskGraph.hasTask(lockTask) && dependenciesLock.exists() &&
                     !project.hasProperty('dependencyLock.ignore')) {
-                logger.info("Using ${dependenciesLock.name} to lock dependencies")
-                def locks = loadLock(dependenciesLock)
-
-                def forcedModules = locks.collect {
-                    overrides.containsKey(it.key) ? "${it.key}:${overrides[it.key]}" : "${it.key}:${it.value.locked}"
-                }
-                logger.debug(forcedModules.toString())
-
-                project.configurations.all {
-                    resolutionStrategy.forcedModules = forcedModules
-                }
+                applyLock(dependenciesLock, overrides)
             } else if (taskGraph.hasTask(lockTask) && !project.hasProperty('dependencyLock.ignore')) {
-                if (project.hasProperty('dependencyLock.overrideFile')) {
-                    logger.info("Using override file ${project['dependencyLock.overrideFile']} to lock dependencies")    
-                }
-                if (project.hasProperty('dependencyLock.override')) {
-                    logger.info("Using command line overrides ${project['dependencyLock.override']}")
-                }
-
-                def overrideModules = overrides.collect { "${it.key}:${it.value}" }
-
-                project.configurations.all {
-                    resolutionStrategy.forcedModules = overrideModules
-                }                        
+                applyOverrides(overrides)
             }
+        }
+    }
+
+    private void configureSaveTask(GenerateLockTask lockTask, DependencyLockExtension extension) {
+        SaveLockTask saveTask = project.tasks.create('saveLock', SaveLockTask)
+        saveTask.conventionMapping.with {
+            generatedLock = { lockTask.dependenciesLock }
+            outputLock = { new File(project.projectDir, extension.lockFile) }
+        }
+        saveTask.dependsOn lockTask
+        saveTask.outputs.upToDateWhen {
+            if (saveTask.generatedLock.exists() && saveTask.outputLock.exists()) {
+                saveTask.generatedLock.text == saveTask.outputLock.text
+            } else {
+                false
+            }
+        }
+    }
+
+    private GenerateLockTask configureLockTask(String clLockFileName, DependencyLockExtension extension, Map overrides) {
+        GenerateLockTask lockTask = project.tasks.create('generateLock', GenerateLockTask)
+        lockTask.conventionMapping.with {
+            dependenciesLock = {
+                new File(project.buildDir, clLockFileName ?: extension.lockFile)
+            }
+            configurationNames = { extension.configurationNames }
+            includeTransitives = { project.hasProperty('dependencyLock.includeTransitives') ? Boolean.parseBoolean(project['dependencyLock.includeTransitives']) : extension.includeTransitives }
+        }
+        lockTask.overrides = overrides
+
+        lockTask
+    }
+
+    void applyOverrides(Map overrides) {
+        if (project.hasProperty('dependencyLock.overrideFile')) {
+            logger.info("Using override file ${project['dependencyLock.overrideFile']} to lock dependencies")
+        }
+        if (project.hasProperty('dependencyLock.override')) {
+            logger.info("Using command line overrides ${project['dependencyLock.override']}")
+        }
+
+        def overrideModules = overrides.collect { "${it.key}:${it.value}" }
+
+        project.configurations.all {
+            resolutionStrategy.forcedModules = overrideModules
+        }
+    }
+
+    void applyLock(File dependenciesLock, Map overrides) {
+        logger.info("Using ${dependenciesLock.name} to lock dependencies")
+        def locks = loadLock(dependenciesLock)
+        def forcedModules = locks.collect {
+            overrides.containsKey(it.key) ? "${it.key}:${overrides[it.key]}" : "${it.key}:${it.value.locked}"
+        }
+        def unusedOverrides = overrides.findAll { !locks.containsKey(it.key) }.collect { "${it.key}:${it.value}" }
+        forcedModules << unusedOverrides
+        logger.debug(forcedModules.toString())
+
+        project.configurations.all {
+            resolutionStrategy.forcedModules = forcedModules
         }
     }
 
     private Map loadOverrides() {
         Map overrides = [:]
+        if (project.hasProperty('dependencyLock.ignore')) {
+            return overrides
+        }
+
         if (project.hasProperty('dependencyLock.overrideFile')) {
-            println project['dependencyLock.overrideFile']
-            File dependenciesLock = new File(project.projectDir, project['dependencyLock.overrideFile'])
-            println dependenciesLock.path
+            File dependenciesLock = new File(project.rootDir, project['dependencyLock.overrideFile'])
             loadLock(dependenciesLock).each { overrides[it.key] = it.value.locked }
             logger.debug "Override file loaded: ${project['dependencyLock.overrideFile']}"
         }
+
         if (project.hasProperty('dependencyLock.override')) {
             project['dependencyLock.override'].tokenize(',').each {
                 def (group, artifact, version) = it.tokenize(':')
-                overrides["${group}:${artifact}"] = version
+                overrides["${group}:${artifact}".toString()] = version
                 logger.debug "Override added for: ${it}"
             }
         }