+++ src/main/groovy/nebula/plugin/dependencylock/tasks/GenerateLockTask.groovy
@@ -21,18 +21,14 @@ import org.gradle.api.artifacts.Dependency
 import org.gradle.api.artifacts.ExternalDependency
 import org.gradle.api.artifacts.ProjectDependency
 import org.gradle.api.artifacts.ResolvedDependency
-import org.gradle.api.tasks.Input
-import org.gradle.api.tasks.OutputFile
 import org.gradle.api.tasks.TaskAction
 
 class GenerateLockTask extends AbstractLockTask {
     String description = 'Create a lock file in build/<configured name>'
-
-    @Input
     Set<String> configurationNames
-
-    @OutputFile
     File dependenciesLock
+    Map overrides
+    Boolean includeTransitives
 
     @TaskAction
     void lock() {
@@ -41,7 +37,7 @@ class GenerateLockTask extends AbstractLockTask {
     }
 
     private readDependenciesFromConfigurations() {
-        def deps = [:].withDefault { [:] }
+        def deps = [:].withDefault { [via: [] as Set, childrenVisited: false] }
         def confs = getConfigurationNames().collect { project.configurations.getByName(it) }
 
         confs.each { Configuration configuration ->
@@ -55,20 +51,62 @@ class GenerateLockTask extends AbstractLockTask {
                 if (!peers.contains(key)) {
                     deps[key.toString()].locked = resolved.moduleVersion
                 }
+                if (getIncludeTransitives()) {
+                    deps[key.toString()].childrenVisited = true
+                    resolved.children.each { handleTransitive(it, deps, key.toString()) }
+                }
+            }
+        }
+
+        getOverrides().each { String key, String overrideVersion ->
+            if (deps.containsKey(key)) {
+                deps[key].viaOverride = overrideVersion
             }
         }
 
         return deps
     }
 
+    private static handleTransitive(ResolvedDependency transitive, Map deps, String parent) {
+        def key = new LockKey(group: transitive.moduleGroup, artifact: transitive.moduleName).toString()
+
+        if (!deps[key].childrenVisited) {
+            deps[key].locked = transitive.moduleVersion
+            deps[key].transitive = true
+            transitive.children.each { handleTransitive(it, deps, key) }
+        }
+        deps[key].via << parent
+    }
+
     private void writeLock(deps) {
-        def strings = deps.collect { k, v -> "  \"${k}\": { \"locked\": \"${v.locked}\", \"requested\": \"${v.requested}\" }"}
+        def strings = deps.collect { String k, Map v -> stringifyLock(k, v) }
         strings = strings.sort()
+        project.buildDir.mkdirs()
         getDependenciesLock().withPrintWriter { out ->
             out.println '{'
             out.println strings.join(',\n')
             out.println '}'
         }
+    }
+
+    private static String stringifyLock(String key, Map lock) {
+        def lockLine = new StringBuilder("  \"${key}\": { \"locked\": \"${lock.locked}\"")
+        if (lock.requested) {
+            lockLine << ", \"requested\": \"${lock.requested}\""
+        }
+        if (lock.transitive) {
+            lockLine << ", \"transitive\": true"
+        }
+        if (lock.viaOverride) {
+            lockLine << ", \"viaOverride\": \"${lock.viaOverride}\""
+        }
+        if (lock.via) {
+            def transitiveFrom = lock.via.sort().collect { "\"${it}\""}.join(', ')
+            lockLine << ", \"via\": [ ${transitiveFrom} ]"
+        }
+        lockLine << ' }'
+
+        return lockLine.toString()
     }
 
     @EqualsAndHashCode