+++ src/test/groovy/nebula/plugin/dependencylock/tasks/GenerateLockTaskSpec.groovy
@@ -15,22 +15,27 @@
  */
 package nebula.plugin.dependencylock.tasks
 
+import nebula.plugin.dependencylock.dependencyfixture.Fixture
 import nebula.test.ProjectSpec
 import org.gradle.testfixtures.ProjectBuilder
 
 class GenerateLockTaskSpec extends ProjectSpec {
     final String taskName = 'generateLock'
 
+    def setupSpec() {
+        Fixture.createFixtureIfNotCreated()
+    }
+
     def 'simple lock'() {
         project.apply plugin: 'java'
 
-        project.repositories { mavenCentral() }
+        project.repositories { maven { url Fixture.repo } }
         project.dependencies {
-            compile 'com.google.guava:guava:14.+'
-            testCompile 'junit:junit:4.+'
+            compile 'test.example:foo:2.+'
+            testCompile 'test.example:baz:1.+'
         }
 
-        GenerateLockTask task = project.tasks.create('lockTestTask', GenerateLockTask)
+        GenerateLockTask task = project.tasks.create(taskName, GenerateLockTask)
         task.dependenciesLock = new File(project.buildDir, 'dependencies.lock')
         task.configurationNames= [ 'testRuntime' ]
 
@@ -40,8 +45,8 @@ class GenerateLockTaskSpec extends ProjectSpec {
         then:
         String lockText = '''\
             {
-              "com.google.guava:guava": { "locked": "14.0.1", "requested": "14.+" },
-              "junit:junit": { "locked": "4.11", "requested": "4.+" }
+              "test.example:baz": { "locked": "1.1.0", "requested": "1.+" },
+              "test.example:foo": { "locked": "2.0.1", "requested": "2.+" }
             }
         '''.stripIndent()
         task.dependenciesLock.text == lockText
@@ -54,18 +59,18 @@ class GenerateLockTaskSpec extends ProjectSpec {
         project.subprojects.add(app)
 
         project.subprojects {
-            repositories { mavenCentral() }
+            repositories { maven { url Fixture.repo } }
         }
 
         common.apply plugin: 'java'
         app.apply plugin: 'java'
         app.dependencies {
             compile app.project(':common')
-            compile 'com.google.guava:guava:14.+'
+            compile 'test.example:foo:2.+'
         }
 
-        GenerateLockTask task = app.tasks.create('lockTestTask', GenerateLockTask)
-        task.dependenciesLock = new File(project.buildDir, 'dependencies.lock')
+        GenerateLockTask task = app.tasks.create(taskName, GenerateLockTask)
+        task.dependenciesLock = new File(app.buildDir, 'dependencies.lock')
         task.configurationNames= [ 'testRuntime' ]
 
         when:
@@ -74,7 +79,117 @@ class GenerateLockTaskSpec extends ProjectSpec {
         then:
         String lockText = '''\
             {
-              "com.google.guava:guava": { "locked": "14.0.1", "requested": "14.+" }
+              "test.example:foo": { "locked": "2.0.1", "requested": "2.+" }
+            }
+        '''.stripIndent()
+        task.dependenciesLock.text == lockText
+    }
+
+    def 'simple transitive lock'() {
+        project.apply plugin: 'java'
+
+        project.repositories { maven { url Fixture.repo } }
+        project.dependencies {
+            compile 'test.example:bar:1.+'
+        }
+
+        GenerateLockTask task = project.tasks.create(taskName, GenerateLockTask)
+        task.dependenciesLock = new File(project.buildDir, 'dependencies.lock')
+        task.configurationNames= [ 'testRuntime' ]
+        task.includeTransitives = true
+
+        when:
+        task.execute()
+
+        then:
+        String lockText = '''\
+            {
+              "test.example:bar": { "locked": "1.1.0", "requested": "1.+" },
+              "test.example:foo": { "locked": "1.0.1", "transitive": true, "via": [ "test.example:bar" ] }
+            }
+        '''.stripIndent()
+        task.dependenciesLock.text == lockText
+    }
+
+    def 'check circular dependency does not loop infinitely'() {
+        project.apply plugin: 'java'
+
+        project.repositories { maven { url Fixture.repo } }
+        project.dependencies {
+            compile 'circular:a:1.+'
+        }
+
+        GenerateLockTask task = project.tasks.create(taskName, GenerateLockTask)
+        task.dependenciesLock = new File(project.buildDir, 'dependencies.lock')
+        task.configurationNames= [ 'testRuntime' ]
+        task.includeTransitives = true
+
+        when:
+        task.execute()
+
+        then:
+        String lockText = '''\
+            {
+              "circular:a": { "locked": "1.0.0", "requested": "1.+", "via": [ "circular:b" ] },
+              "circular:b": { "locked": "1.0.0", "transitive": true, "via": [ "circular:a" ] }
+            }
+        '''.stripIndent()
+        task.dependenciesLock.text == lockText
+    }
+
+    def 'one level transitive test'() {
+        project.apply plugin: 'java'
+
+        project.repositories { maven { url Fixture.repo } }
+        project.dependencies {
+            compile 'test.example:bar:1.+'
+            compile 'test.example:foobaz:1.+'
+        }
+
+        GenerateLockTask task = project.tasks.create(taskName, GenerateLockTask)
+        task.dependenciesLock = new File(project.buildDir, 'dependencies.lock')
+        task.configurationNames= [ 'testRuntime' ]
+        task.includeTransitives = true
+
+        when:
+        task.execute()
+
+        then:
+        String lockText = '''\
+            {
+              "test.example:bar": { "locked": "1.1.0", "requested": "1.+" },
+              "test.example:baz": { "locked": "1.0.0", "transitive": true, "via": [ "test.example:foobaz" ] },
+              "test.example:foo": { "locked": "1.0.1", "transitive": true, "via": [ "test.example:bar", "test.example:foobaz" ] },
+              "test.example:foobaz": { "locked": "1.0.0", "requested": "1.+" }
+            }
+        '''.stripIndent()
+        task.dependenciesLock.text == lockText
+    }
+
+    def 'multi-level transitive test'() {
+        project.apply plugin: 'java'
+
+        project.repositories { maven { url Fixture.repo } }
+        project.dependencies {
+            compile 'test.example:transitive:1.0.0'
+        }
+
+        GenerateLockTask task = project.tasks.create(taskName, GenerateLockTask)
+        task.dependenciesLock = new File(project.buildDir, 'dependencies.lock')
+        task.configurationNames= [ 'testRuntime' ]
+        task.includeTransitives = true
+
+        when:
+        task.execute()
+
+        then:
+        String lockText = '''\
+            {
+              "test.example:bar": { "locked": "1.0.0", "transitive": true, "via": [ "test.example:transitive" ] },
+              "test.example:baz": { "locked": "1.0.0", "transitive": true, "via": [ "test.example:foobaz" ] },
+              "test.example:foo": { "locked": "1.0.1", "transitive": true, "via": [ "test.example:bar", "test.example:foobaz" ] },
+              "test.example:foobaz": { "locked": "1.0.0", "transitive": true, "via": [ "test.example:transitive" ] },
+              "test.example:transitive": { "locked": "1.0.0", "requested": "1.0.0" }
             }
         '''.stripIndent()
         task.dependenciesLock.text == lockText